Enable more than 2 groups of keywords.
Before this revision, only two groups of keywords are allowed in data collection. The keywords in a group are synonyms and will be "OR" together, while the groups will be "AND" together to search the papers on OpenAlex. This revision removes the restriction of the number of groups. Now we can use more than two groups of keywords. That means we can search for papers with more precision.